+++ src/gui/headers/LeftSidebarWidget.h
@@ -2,16 +2,15 @@
 
 #include "widget.h"
 #include "EntityControlWidget.h"
-//#include "../../graphics/GraphicsController.h"
 #include <vector>
 #include <string>
-#include <unordered_map>>
+#include <unordered_map>
 
 class LeftSidebarWidget : public Widget
 {
 private:
 	std::vector<std::string> nodes;
-	unsigned int selectedIndex;
+	size_t selectedIndex;
 
 public:
 	LeftSidebarWidget();
